loft-partitions/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
redis==4.5.4
celery==5.3.4
python-dotenv==1.0.0
django-cors-headers==4.3.1
Pillow==10.0.0
django-filter==23.3
python-dotenv==1.0.0